backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
openai==1.3.0
pgvector==0.2.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6